Apply #2804
git commit -m
git commit -m Apply

+++ Application/DTOs/Payment/PaymentDTO.cs
@@ -11,9 +11,9 @@ namespace TouRest.Application.DTOs.Payment
         public Guid Id { get; set; }
         public Guid BookingId { get; set; }
         public long OrderCode { get; set; }
-        public decimal Amount { get; set; }
-        public decimal DiscountAmount { get; set; }
-        public decimal FinalAmount { get; set; }
+        public long Amount { get; set; }
+        public long DiscountAmount { get; set; }
+        public long FinalAmount { get; set; }
         public string Status { get; set; } = null!;
         public string? CheckoutUrl { get; set; }
         public DateTime ExpiredAt { get; set; }
