+++ Application/DTOs/Refund/RefundDTO.cs
@@ -11,7 +11,7 @@ namespace TouRest.Application.DTOs.Refund
         public Guid Id { get; set; }
         public Guid BookingId { get; set; }
         public Guid PaymentId { get; set; }
-        public decimal TotalRefundAmount { get; set; }
+        public long TotalRefundAmount { get; set; }
         public string InitiatedBy { get; set; } = null!;
         public string? Reason { get; set; }
         public string? AdminNote { get; set; }
